requirements.txt
Flask
supabase
gunicorn
python-dotenv
Flask-CORS